+++ packages/block-editor/src/components/global-styles/hooks.js
@@ -8,7 +8,7 @@ import fastDeepEqual from 'fast-deep-equal/es6';
  */
 import { useContext, useCallback, useMemo } from '@wordpress/element';
 import { useSelect } from '@wordpress/data';
-import { store as blocksStore } from '@wordpress/blocks';
+import { store as blocksStore, hasBlockSupport } from '@wordpress/blocks';
 import { _x } from '@wordpress/i18n';
 
 /**
@@ -19,10 +19,11 @@ import { getValueFromObjectPath, setImmutably } from '../../utils/object';
 import { GlobalStylesContext } from './context';
 import { unlock } from '../../lock-unlock';
 
-const EMPTY_CONFIG = { settings: {}, styles: {} };
+const EMPTY_CONFIG = { settings: {}, styles: {}, behaviors: {} };
 
 const VALID_SETTINGS = [
 	'appearanceTools',
+	'behaviors',
 	'useRootPaddingAwareAlignments',
 	'border.color',
 	'border.radius',
@@ -88,7 +89,6 @@ export const useGlobalStylesReset = () => {
 
 export function useGlobalSetting( propertyPath, blockName, source = 'all' ) {
 	const { setUserConfig, ...configs } = useContext( GlobalStylesContext );
-
 	const appendedBlockPath = blockName ? '.blocks.' + blockName : '';
 	const appendedPropertyPath = propertyPath ? '.' + propertyPath : '';
 	const contextualPath = `settings${ appendedBlockPath }${ appendedPropertyPath }`;
@@ -135,7 +135,6 @@ export function useGlobalSetting( propertyPath, blockName, source = 'all' ) {
 			setImmutably( currentConfig, contextualPath.split( '.' ), newValue )
 		);
 	};
-
 	return [ settingValue, setSetting ];
 }
 
@@ -327,7 +326,8 @@ export function useSettingsForBlockElement(
 			const sides = Array.isArray( supports?.spacing?.[ key ] )
 				? supports?.spacing?.[ key ]
 				: supports?.spacing?.[ key ]?.sides;
-			if ( sides?.length ) {
+			// Check if spacing type is supported before adding sides.
+			if ( sides?.length && updatedSettings.spacing?.[ key ] ) {
 				updatedSettings.spacing = {
 					...updatedSettings.spacing,
 					[ key ]: {
@@ -460,4 +460,113 @@ export function useGradientsPerOrigin( settings ) {
 		defaultGradients,
 		shouldDisplayDefaultGradients,
 	] );
+}
+
+export function __experimentalUseGlobalBehaviors( blockName, source = 'all' ) {
+	const {
+		merged: mergedConfig,
+		base: baseConfig,
+		user: userConfig,
+		setUserConfig,
+	} = useContext( GlobalStylesContext );
+	const finalPath = ! blockName
+		? `behaviors`
+		: `behaviors.blocks.${ blockName }`;
+
+	let rawResult, result;
+	switch ( source ) {
+		case 'all':
+			rawResult = getValueFromObjectPath( mergedConfig, finalPath );
+			result = getValueFromVariable( mergedConfig, blockName, rawResult );
+			break;
+		case 'user':
+			rawResult = getValueFromObjectPath( userConfig, finalPath );
+			result = getValueFromVariable( mergedConfig, blockName, rawResult );
+			break;
+		case 'base':
+			rawResult = getValueFromObjectPath( baseConfig, finalPath );
+			result = getValueFromVariable( baseConfig, blockName, rawResult );
+			break;
+		default:
+			throw 'Unsupported source';
+	}
+
+	const animation = result?.lightbox?.animation || 'zoom';
+
+	const setBehavior = ( newValue ) => {
+		let newBehavior;
+		// The user saves with Apply Globally option.
+		if ( typeof newValue === 'object' ) {
+			newBehavior = newValue;
+		} else {
+			switch ( newValue ) {
+				case 'lightbox':
+					newBehavior = {
+						lightbox: {
+							enabled: true,
+							animation,
+						},
+					};
+					break;
+				case 'fade':
+					newBehavior = {
+						lightbox: {
+							enabled: true,
+							animation: 'fade',
+						},
+					};
+					break;
+				case 'zoom':
+					newBehavior = {
+						lightbox: {
+							enabled: true,
+							animation: 'zoom',
+						},
+					};
+					break;
+				case '':
+					newBehavior = {
+						lightbox: {
+							enabled: false,
+							animation,
+						},
+					};
+					break;
+				default:
+					break;
+			}
+		}
+		setUserConfig( ( currentConfig ) =>
+			setImmutably( currentConfig, finalPath.split( '.' ), newBehavior )
+		);
+	};
+	let behavior = '';
+	if ( result === undefined ) behavior = 'default';
+	if ( result?.lightbox.enabled ) behavior = 'lightbox';
+
+	return { behavior, inheritedBehaviors: result, setBehavior };
+}
+
+export function __experimentalUseHasBehaviorsPanel(
+	settings,
+	name,
+	{ blockSupportOnly = false } = {}
+) {
+	if ( ! settings?.behaviors ) {
+		return false;
+	}
+
+	// If every behavior is disabled on block supports, do not show the behaviors inspector control.
+	const hasSomeBlockSupport = Object.keys( settings?.behaviors ).some(
+		( key ) => hasBlockSupport( name, `behaviors.${ key }` )
+	);
+
+	if ( blockSupportOnly ) {
+		return hasSomeBlockSupport;
+	}
+
+	// If every behavior is disabled, do not show the behaviors inspector control.
+	return Object.values( settings?.behaviors ).some(
+		( value ) => value === true && hasSomeBlockSupport
+	);
 }